+++ src/Microsoft.Azure.ServiceBus/Amqp/AmqpTopicClient.cs
@@ -1,6 +1,8 @@
 ﻿// Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
+using Microsoft.Azure.ServiceBus.Core;
+
 namespace Microsoft.Azure.ServiceBus.Amqp
 {
     using System.Threading.Tasks;
