+++ src/Microsoft.Azure.ServiceBus/Core/IMessageSender.cs
@@ -1,7 +1,7 @@
 ﻿// Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-namespace Microsoft.Azure.ServiceBus
+namespace Microsoft.Azure.ServiceBus.Core
 {
     using System;
     using System.Collections.Generic;
@@ -11,7 +11,7 @@ namespace Microsoft.Azure.ServiceBus
     {
         Task SendAsync(BrokeredMessage brokeredMessage);
 
-        Task SendAsync(IEnumerable<BrokeredMessage> brokeredMessages);
+        Task SendAsync(IList<BrokeredMessage> brokeredMessages);
 
         Task<long> ScheduleMessageAsync(BrokeredMessage message, DateTimeOffset scheduleEnqueueTimeUtc);
