+++ src/Microsoft.Azure.ServiceBus/Core/IMessageSessionEntity.cs
@@ -1,7 +1,7 @@
 ﻿// Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-namespace Microsoft.Azure.ServiceBus
+namespace Microsoft.Azure.ServiceBus.Core
 {
     using System;
     using System.Threading.Tasks;
