+++ src/Microsoft.Azure.ServiceBus/TopicClient.cs
@@ -1,6 +1,8 @@
 ﻿// Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
+using Microsoft.Azure.ServiceBus.Core;
+
 namespace Microsoft.Azure.ServiceBus
 {
     using System;
@@ -59,7 +61,7 @@ namespace Microsoft.Azure.ServiceBus
             return this.SendAsync(new[] { brokeredMessage });
         }
 
-        public Task SendAsync(IEnumerable<BrokeredMessage> brokeredMessages)
+        public Task SendAsync(IList<BrokeredMessage> brokeredMessages)
         {
             return this.InnerSender.SendAsync(brokeredMessages);
         }
